+++ packages/plugins/prometheus/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @envelop/prometheus
 
+## 6.3.1
+
+### Patch Changes
+
+- Updated dependencies [d5c2c9a]
+  - @envelop/core@2.3.1
+
+## 6.3.0
+
+### Minor Changes
+
+- Updated dependencies [af23408]
+  - @envelop/core@2.3.0
+
 ## 6.2.0
 
 ### Minor Changes