+++ packages/plugins/response-cache-redis/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @envelop/response-cache-redis
 
+## 0.4.1
+
+### Patch Changes
+
+- @envelop/response-cache@2.3.1
+
+## 0.4.0
+
+### Patch Changes
+
+- @envelop/response-cache@2.3.0
+
 ## 0.3.0
 
 ### Patch Changes