+++ packages/plugins/sentry/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @envelop/sentry
 
+## 3.4.0
+
+### Minor Changes
+
+- 0be88fe: Expose sentry event id in error extensions
+
+### Patch Changes
+
+- Updated dependencies [d5c2c9a]
+  - @envelop/core@2.3.1
+
+## 3.3.0
+
+### Minor Changes
+
+- Updated dependencies [af23408]
+  - @envelop/core@2.3.0
+
 ## 3.2.0
 
 ### Minor Changes