+++ packages/types/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @envelop/types
 
+## 2.2.0
+
+### Minor Changes
+
+- af23408: Pass context to registerContextErrorHandler
+
+  ```ts
+  export const useMyHook = (): Plugin => {
+    return {
+      onPluginInit(context) {
+        context.registerContextErrorHandler(({ context }) => {
+          console.error('Error occurred during context creation but at least I have the  context so far', context);
+        });
+      },
+    };
+  };
+  ```
+
 ## 2.1.0
 
 ### Minor Changes